First step to refactoring 'PayloadTransport' into 'ViewController'
