+++ packages/@magic-sdk/provider/src/README.md
@@ -16,7 +16,7 @@ There are numerous code paths that [`./modules`](./modules) can follow to intera
 // `magic.auth` and `magic.user` methods follow this pattern.
 magic.user.getMetadata();
 
-BaseModule.request -> PayloadTransport.post -> Window.postMessage -> iframe
+BaseModule.request -> ViewController.post -> Window.postMessage -> iframe
 ```
 
 ### Case #2: Web3 `>=1.0.0-beta.38` using `send`
@@ -26,7 +26,7 @@ This represents the most current behavior of Web3:
 ```ts
 web3.eth.sendTransaction(...);
 
-RPCProviderModule.send -> BaseModule.request -> PayloadTransport.post -> Window.postMessage -> iframe
+RPCProviderModule.send -> BaseModule.request -> ViewController.post -> Window.postMessage -> iframe
 ```
 
 ### Case #3: Web3 `<=1.0.0-beta.37` using `send`
@@ -36,7 +36,7 @@ This is an edge case affecting a small subset of Web3 pre-`1.x` beta versions:
 ```ts
 web3.eth.sendTransaction(...);
 
-RPCProviderModule.send -> RPCProviderModule.sendAsync -> PayloadTransport.post -> Window.postMessage -> iframe
+RPCProviderModule.send -> RPCProviderModule.sendAsync -> ViewController.post -> Window.postMessage -> iframe
 ```
 
 ### Case #4: Legacy Web3 using `sendAsync`
@@ -46,7 +46,7 @@ This is the legacy behavior of Web3 providers, though it is still widely support
 ```ts
 web3.eth.sendTransaction(...);
 
-RPCProviderModule.sendAsync -> PayloadTransport.post -> Window.postMessage -> iframe
+RPCProviderModule.sendAsync -> ViewController.post -> Window.postMessage -> iframe
 ```
 
 ## React Native
