+++ packages/@magic-sdk/provider/src/index.ts
@@ -2,7 +2,6 @@
 
 export { SDKBase, MagicSDKAdditionalConfiguration } from './core/sdk';
 export { createSDK } from './core/sdk-environment';
-export { PayloadTransport } from './core/payload-transport';
 export { ViewController } from './core/view-controller';
 export * from './core/sdk-exceptions';
 export { Extension } from './modules/base-extension';
