+++ packages/@magic-sdk/provider/test/constants.ts
@@ -1,5 +1,3 @@
-/* eslint-disable no-underscore-dangle */
-
 export const MAGIC_RELAYER_FULL_URL = 'https://auth.magic.link';
 export const TEST_API_KEY = 'pk_test_123';
 export const LIVE_API_KEY = 'pk_live_123';
