+++ packages/@magic-sdk/provider/test/spec/core/json-rpc/json-rpc-response/constructor.spec.ts
@@ -1,5 +1,3 @@
-/* eslint-disable no-underscore-dangle */
-
 import browserEnv from '@ikscodes/browser-env';
 import { JsonRpcResponse } from '../../../../../src/core/json-rpc';
