+++ packages/@magic-sdk/provider/test/spec/core/sdk-exceptions/magic-rpc-error/constructor.spec.ts
@@ -1,5 +1,3 @@
-/* eslint-disable no-underscore-dangle */
-
 import browserEnv from '@ikscodes/browser-env';
 import { MagicRPCError } from '../../../../../src/core/sdk-exceptions';
