+++ packages/@magic-sdk/types/src/core/i18n.ts
@@ -1,0 +1,41 @@
+export type SupportedLocale =
+  | 'en_US'
+  | 'pl_PL'
+  | 'af'
+  | 'az'
+  | 'ca'
+  | 'cy'
+  | 'da'
+  | 'de'
+  | 'et'
+  | 'en'
+  | 'es'
+  | 'fr'
+  | 'hr'
+  | 'id'
+  | 'it'
+  | 'lv'
+  | 'lt'
+  | 'hu'
+  | 'nl'
+  | 'no'
+  | 'pl'
+  | 'pt'
+  | 'ro'
+  | 'sk'
+  | 'sl'
+  | 'fi'
+  | 'sv'
+  | 'tr'
+  | 'vi'
+  | 'cs'
+  | 'el'
+  | 'bg'
+  | 'mk'
+  | 'ru'
+  | 'sr'
+  | 'th'
+  | 'zh_TW'
+  | 'zh_CN'
+  | 'ja'
+  | 'ko';
