+++ packages/@magic-sdk/types/src/index.ts
@@ -1,10 +1,11 @@
 // Only re-export types that are intended for the public API from this file.
 
+export * from './core/exception-types';
+export * from './core/i18n';
 export * from './core/json-rpc-types';
 export * from './core/message-types';
-export * from './core/exception-types';
 export * from './core/query-types';
 
 export * from './modules/auth-types';
-export * from './modules/user-types';
 export * from './modules/rpc-provider-types';
+export * from './modules/user-types';
