+++ packages/magic-sdk/src/index.cdn.ts
@@ -1,4 +1,3 @@
-/* eslint-disable no-underscore-dangle */
 /* istanbul ignore file */
 
 import 'regenerator-runtime/runtime';
@@ -8,7 +7,6 @@ import * as publicAPI from '@magic-sdk/commons';
 import localForage from 'localforage';
 import memoryDriver from 'localforage-driver-memory';
 import { IframeController } from './iframe-controller';
-import { WebTransport } from './web-transport';
 
 const Magic = createSDK(SDKBase, {
   platform: 'web',
@@ -16,7 +14,6 @@ const Magic = createSDK(SDKBase, {
   version: '%WEB_VERSION%',
   defaultEndpoint: 'https://auth.magic.link/',
   ViewController: IframeController,
-  PayloadTransport: WebTransport,
   configureStorage: /* istanbul ignore next */ async () => {
     const lf = localForage.createInstance({
       name: 'MagicAuthLocalStorageDB',
