+++ packages/magic-sdk/src/index.ts
@@ -1,10 +1,7 @@
-/* eslint-disable no-underscore-dangle */
-
 import { SDKBase, createSDK } from '@magic-sdk/provider';
 import localForage from 'localforage';
 import * as memoryDriver from 'localforage-driver-memory';
 import { IframeController } from './iframe-controller';
-import { WebTransport } from './web-transport';
 
 export * from '@magic-sdk/commons';
 
@@ -14,7 +11,6 @@ export const Magic = createSDK(SDKBase, {
   version: '%WEB_VERSION%',
   defaultEndpoint: 'https://auth.magic.link/',
   ViewController: IframeController,
-  PayloadTransport: WebTransport,
   configureStorage: /* istanbul ignore next */ async () => {
     const lf = localForage.createInstance({
       name: 'MagicAuthLocalStorageDB',
